Moved ProgressBarCallback into callbacks.py
